remove go tests
go tests were using the websocket payloads to test everything. Now ws payloads are in the form of html so it makes no sense to test them in go. We can just test with the e2e suite.

What it could make sense is to write some unit tests for the business folder.